+++ CHANGELOG.md
@@ -22,7 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Kotlin `1.8.10`
 - Jetpack Compose Compiler to `1.4.3`
 - Jetpack Compose Runtime to `1.3.3`
-- Jetpack Compose UI to `1.3.1`
+- Jetpack Compose UI to `1.4.0-beta02`
 - Jetpack Compose Material 3 to `1.1.0-alpha07`
 - Ktlint to `11.2.0`
 - Androidx Lifecycle to `2.6.0-rc01`